requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosmtplib==4.0.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asyncer==0.0.7
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
certifi==2025.1.31
chainlit==2.5.5
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.52
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
Lazify==0.4.0
literalai==0.1.201
lxml==4.9.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.6.0
monotonic==1.6
multidict==6.4.3
mypy-extensions==1.0.0
neo4j==5.28.1
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.74.1
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-alephalpha==0.39.1
opentelemetry-instrumentation-anthropic==0.39.1
opentelemetry-instrumentation-bedrock==0.39.1
opentelemetry-instrumentation-chromadb==0.39.1
opentelemetry-instrumentation-cohere==0.39.1
opentelemetry-instrumentation-crewai==0.39.1
opentelemetry-instrumentation-google-generativeai==0.39.1
opentelemetry-instrumentation-groq==0.39.1
opentelemetry-instrumentation-haystack==0.39.1
opentelemetry-instrumentation-lancedb==0.39.1
opentelemetry-instrumentation-langchain==0.39.1
opentelemetry-instrumentation-llamaindex==0.39.1
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-instrumentation-marqo==0.39.1
opentelemetry-instrumentation-milvus==0.39.1
opentelemetry-instrumentation-mistralai==0.39.1
opentelemetry-instrumentation-ollama==0.39.1
opentelemetry-instrumentation-openai==0.39.1
opentelemetry-instrumentation-pinecone==0.39.1
opentelemetry-instrumentation-qdrant==0.39.1
opentelemetry-instrumentation-replicate==0.39.1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-sagemaker==0.39.1
opentelemetry-instrumentation-sqlalchemy==0.52b1
opentelemetry-instrumentation-threading==0.52b1
opentelemetry-instrumentation-together==0.39.1
opentelemetry-instrumentation-transformers==0.39.1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-vertexai==0.39.1
opentelemetry-instrumentation-watsonx==0.39.1
opentelemetry-instrumentation-weaviate==0.39.1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-semantic-conventions-ai==0.4.3
opentelemetry-util-http==0.52b1
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
posthog==3.25.0
propcache==0.3.1
protobuf==5.29.4
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-futures==1.0.2
requests-toolbelt==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.41.3
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.39.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uptrace==1.31.0
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yahooquery==2.3.7
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
